Initial commit of podlabel
This script is to be attached on the cube transform of the pod.
Under the current implementation that might not be too necessary, but if
we ever decide to switch over to OnWillRenderObject implementation, this
script has to be on the cube since that is the only part of the pod seen
by the camera.

Will work on making the GUIText look nicer with boxes for my next
commit.